requirements.txt
flask
gunicorn
opencv-contrib-python-headless
numpy==1.16.5
tensorflow==1.13.1
keras==2.2.4
flask-login==0.5.0
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.6
psycopg2==2.8.6
pandas
pillow